Refactor iconic taxon to be inside Confetti component
